Move make_zipfile into utils and stop writing folders
